* Day 13 Part 2 runs buttery smooth after cleaning up (<0ms)!!

+++ AdventOfCode2020/Day13.h
@@ -17,7 +17,7 @@ private:
 	int _target = 0, _earliest = INT_MAX, _busId = 0;
 	vector<int> _busses;
 	map<int, int> _offsets;
-	map<int, pair<int, int>> _test;
+	map<int, pair<int, long long>> _test;
 	long long _startAt = 0;
 public:
 	void ParseInput();
